TarasciiMadness v1.4 - full codebase update
Brings the repository up to date with the latest TarasciiMadness
development on the current OpenJK base. Includes barrel placement,
team mechanics, win/loss celebration, scoreboard overhaul, HUD,
admin commands, disruptor weapon mechanics, tmSyncCvar system,
client plugin v1.4 enforcement, CI/CD build pipeline, and assets.

+++ codemp/qcommon/cm_local.h
@@ -45,10 +45,10 @@ typedef struct cLeaf_s {
 	int			cluster;
 	int			area;
 
-	ptrdiff_t	firstLeafBrush;
+	int			firstLeafBrush;
 	int			numLeafBrushes;
 
-	ptrdiff_t	firstLeafSurface;
+	int			firstLeafSurface;
 	int			numLeafSurfaces;
 } cLeaf_t;
